Bug fixes and refactoring on CMFW
